+++ examples/webpage/src/pages/docs/doc_pages/mod.rs
@@ -1,7 +1,7 @@
 pub mod animations;
 pub mod colors;
 pub mod doc_page;
-pub mod examples;
 pub mod flex;
+pub mod getting_started;
 pub mod pseudoclass_themes;
 pub mod responsive_design;